SHIRO-141:
Refactored SecurityManager.createSubject and SubjectFactory.createSubject methods to take in a new interface: SubjectContext.  SubjectContext is essentially a Map with a few extra type-safe getters/setters and some resolve* methods.  The resolve methods use heuristics to acquire data used to construct subject instances. 

This logic used to be entirely in the SubjectFactory implementations, but it was also needed by the RememberMeManager implementations, which didn't have access to/knowledge of a SubjectFactory (and shouldn't).  So that logic was extracted to a SubjectContext implementation that could be shared across both RememberMeManager and SubjectFactory implementations.

Also removed two Exceptions that were never used.  Updated test cases to reflect the new interface addition.

Finally, the default CipherService used by the AbstractRememberMeManager is now an AesCipherService instead of the previous BlowfishCipherService default.  AES is more commonly used in most end-user environments.  Existing RememberMe cookies will therefore fail with this configuration, but this is a non-critical problem - the next login w/ rememberMe enabled will create an AES-encrypted cookie and work from that point on.

git-svn-id: https://svn.apache.org/repos/asf/incubator/shiro/trunk@937094 13f79535-47bb-0310-9956-ffa450edef68