SHIRO-128:
Added new Cookie interface and SimpleCookie implementation.  Created new CookieRememberMeManager implementation that did the same thing as the old WebRememberMeManager but with easier-to-understand Cookie semantics instead of the obscure and confusing WebAttribute concept.

The addition of this interface and simple implementation allowed the removal of all the convenience methods related to 1) remember me cookies and 2) session id cookies

The new Cookie concept can be configured like any other pojo in INI configuration (or any other compatible JavaBeans config mechanism)

This accounts for almost half of the convenience methods to remove.   Other half still needs to be completed.

git-svn-id: https://svn.apache.org/repos/asf/incubator/shiro/trunk@937127 13f79535-47bb-0310-9956-ffa450edef68