Include ISIS Security extensions to Shiro for RBAC and SSO
The Apache ISIS project made a very interesting extension to Shiro
allowing the automatic fetch of LDAP Groups and assigning them 
to roles in order to resolve the user’s permissions.

This extension makes sense *inside* Shiro as it can be reused in many
other context even outside ISIS.

Additionally I have extended the LDAP Realm and Authentication Filter
in order to allow SSO with an external authentication system resolved
through an HTTP Proxy (e.g. Kerberos, but it could be anything else, 
including commercial products such as Site-Minder) so that they can
be used in conjunction.