Update the webstart-maven-plugin to the latest release
This requires some changes to the configuration:
* <pack200> moved to <pack200><enabled>
* the keystore path is (now?) relative to the project directory
* the keytool dependencies seem to be missing

This looks partially like issues in the webstart plugin, which
unfortunately is in the codehaus->github migration bubble.

With this change shiro builds with both the toolchain configuration
and JAVA_HOME (due to the webstart plugin ignoring toolchains, presumably)
pointing to a Java 7 JDK.

This makes it possible to compile shiro with Java 8 and up.