[SHIRO-457] Set SecurityManager in createSubject()
When one does not use a static SecurityManager an exception is thrown
every time a new (unauthenticated) subject successfully logs in. See
the associated JIRA issue:

    https://issues.apache.org/jira/browse/SHIRO-457

This commit ensures Subject.login() will first use the SecurityManager
passed to its Subject.Builder (if one was specified) instead of
immediately starting with the one found in the ThreadContext or
statically in SecurityUtils.