Implemented "Replace Conditional with Polymorphism" refactoring
The switch statement in the original code chooses a different course of action depending on the value of the system property INTEGRATION TEST MODE PROPERTY. This method can lead to maintenance problems, such as when new cases are introduced or when an old case needs to have its behaviour changed. The switch statement may also make the code more challenging to comprehend.

A hierarchy of subclasses that correspond to each situation of the switch expression can be produced by restructuring the code using polymorphism. The code is easier to understand since each subclass has functionality that is unique to that subclass's particular case. The usage of polymorphism also improves the code's extensibility by allowing new scenarios to be introduced without changing the original code by adding new subclasses.