Implemented Replace Conditional with Polymorphism
The original code has a switch statement that checks the value of a string parameter and sets the type variable to a constant based on that value. Then, it creates an instance of Argon2Parameters using the value of type and executes the generateBytes method of Argon2BytesGenerator to generate a hash. Finally, it returns a new Argon2Hash object containing the hash and other parameters.

By replacing the switch statement with polymorphism, a common interface can be defined for all possible implementations of the generateHash method, which is responsible for generating the hash. Each subclass of Argon2Algorithm can implement the generateHash method according to the corresponding value of the string parameter, thus encapsulating the algorithm-specific logic in a single class. This way, new algorithms can be easily added by adding new subclasses without modifying the existing code.